cd-code/requirements.txt
aws-cdk-lib==2.100.0
constructs==10.0.0
